+++ api/api.ts
@@ -1,8 +1,19 @@
 import axios from 'axios';
 
 const axiosInstance = axios.create({
-  baseURL: process.env.REACT_APP_API_BASE_URL,
+  baseURL: process.env.NEXT_PUBLIC_API_BASE_URL,
   headers: { 'Content-Type': 'application/json' },
+  withCredentials: true,
+});
+
+// 요청마다 `Authorization` 헤더를 동적으로 추가 (인터셉트 활용)
+// ~로그인에선 필요없는 부분~
+axiosInstance.interceptors.request.use((config) => {
+  const token = localStorage.getItem('accessToken');
+  if (token) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+  return config;
 });
 
 export default axiosInstance;
